[Improve] [Core] Extract the CommandArgs implement class field (#1720)
* Extract the CommandArgs implement class field
* rename CommandArgs to AbstractCommandArgs
